Add failing test
I want to make sure errors pass through from the contract that inherits Delegatable.
This has turned into a problem for MobyMask, because it is kinda inexplicably failing suddenly.

The strange thing now is that in addition to the first (new) test failing (good), two unrelated tests are suddenly failing.
These two tests are expecting specific typedDataHash values, which looks a bit brittle, but I don't understand why they would not be deterministic.